pyproject.toml
[project]
name = "alm-system"
version = "0.1.0"
requires-python = ">=3.12.11"
dependencies = [
    "anthropic>=0.86.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "jax>=0.9.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.30.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "pyqt6>=6.11.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.32.5",
    "rq>=2.7.0",
    "scipy>=1.17.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "tests.sample_data.*",
]
ignore_errors = true

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "jupyter>=1.0",
    "fakeredis>=2.34.1",
    "responses>=0.26.0",
    "types-requests>=2.33.0.20260408",
    "types-pyyaml>=6.0.12.20260408",
    "types-redis>=4.6.0.20241004",
    "mypy>=1.20.1",
]
